Add support for app code authentication

+++ phpipamsdk/phpipam.py
@@ -27,16 +27,24 @@ class PhpIpamApi(object):
     _api_uri = ''
     _api_token = ''
     _api_verify_ssl = False
+    _api_appcode_auth = False
     _api_headers = {
         'accept': 'application/json',
         'content-type': 'application/json'
     }
 
-    def __init__(self, api_uri=None, api_verify_ssl=None):
+    def __init__(self, api_uri=None, api_appcode=None, api_verify_ssl=None):
         if api_uri is None:
             self._api_uri = Configuration().api_uri
         else:
             self._api_uri = api_uri
+        if api_appcode is None:
+            if Configuration().api_appcode:
+                self._api_token = Configuration().api_appcode
+                self._api_appcode_auth = True
+        else:
+            self._api_token = api_appcode
+            self._api_appcode_auth = True
         if api_verify_ssl is None:
             self._api_verify_ssl = Configuration().api_verify_ssl
         else:
@@ -69,6 +77,9 @@ class PhpIpamApi(object):
 
     def login(self, auth=None):
         """ authenticate to API """
+        if self._api_appcode_auth:
+            self._api_headers['phpipam-token'] = self._api_token
+            return
         if auth is None:
             auth = (Configuration().api_username, Configuration().api_password)
         result = self.api_send_request(path='user/', auth=auth, method='post')
@@ -77,15 +88,21 @@ class PhpIpamApi(object):
 
     def get_token(self):
         """ get auth token """
+        if self._api_appcode_auth:
+            return
         uri = 'user/'
         self.api_send_request(path=uri, method='get')
 
     def refresh_token(self):
         """ refresh auth token """
+        if self._api_appcode_auth:
+            return
         uri = 'user/'
         self.api_send_request(path=uri, method='patch')
 
     def logout(self):
         """ delete session """
+        if self._api_appcode_auth:
+            return
         uri = 'user/'
         self.api_send_request(path=uri, method='delete')
